Fix Nodes routes
