+++ src/styles/components/tabs/styles.less
@@ -6,6 +6,7 @@
     display: flex;
     margin-left: 0;
     margin-top: 0;
+    width: 100%;
 
     &.tall {
       margin-top: @base-spacing-unit;
@@ -89,7 +90,7 @@
         }
 
         &.active {
-          color: @white;
+          color: @purple;
         }
       }
     }
